Implement format on save
Unfortunately there doesn't seem to be a unified way to see if we're
dealing with a Ruby file. Package activation is contingent
on the Ruby or RoR package being loaded. The keyboard shortcut is done
by CSS selectors on the file grammar. And finally the on save check
is a plain JS conditional.